Avoid going back to WindowCache on already valid WindowCursors
Its not uncommon for a WindowCursor to be used to read data from the
pack file when it is already holding the correct window handle.  It
happens a lot on delta compressed objects as we need to do a second
read to locate the delta base information.  Letting the cursor check
to see if it is valid before it goes back into the cache saves us a
lock acquire as well as cache lookup costs, which can reduce our
total CPU time.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>