Hold onto pack windows while decompressing objects
Typically when we build a PackedObjectLoader the caller is going to
immediately turn around and pull the bytes for the object pack out
of the very same packfile.  In such a case we can recycle the window
cursor we used to read the object header and delta base information
as it is positioned to cover the object's compressed data.  This can
reduce locking and window lookup overheads associated with the data
decompression part of object loading, further improving our access.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>