Allow RevWalk to recycle the WindowCursor during walking
By hanging on to the window cursor while we are walking through
the revision history we can get better access rates to the raw
pack file data as multiple commits usually fit inside of one of
our windows and commit locality within the pack file often  means
the parents of a given commit are immediately behind it, and thus
part of the same window as the last commit we parsed.

We also can hang onto a temporary byte buffer for object header
parsing, as the GC does not need to allocate this buffer each
time we lookup something from a pack file.  In theory this is an
object that should be going onto the call stack and not into the
heap, but current JVMs are not likely to be able to see this due
to how many methods we pass the byte[] into.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>