Make "when" part of a PersonIdent required so we stop boxing values
We always have a time part for a person identity, except for when
we are looking at a corrupt commit.  The corrupt commit case never
happens, and if/when it does a time value of 0 (meaning the epoch)
is just as acceptable as a null time value.  It actually simplifies
a lot of code as we don't have to worry about getting a null value
back from the getWhen() method.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>