Add utility methods for decoding the author, committer to RevCommit
Some applications will find it useful to parse the person identity
out of a RevCommit, without going through the slower Commit class.
Eventually we should be able to remove the Commit class entirely,
as RevCommit is growing to the point where it does not need to use
the Commit class at all.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>