Allow RevCommit subclasses to implement their own reset logic
If an application extends RevCommit they may have done so to store
additional state information within the commit object.  If they do
a reset on their RevWalk to clear the commit state we need to let
the application also clear the state of any commit we visited under
a prior walk.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>