Define RevObjectList, RevCommitList as efficient append-only Lists
We use a block allocation strategy to incrementally grow our lists
as new items are inserted onto the end of the list.  This design is
a reasonable trade-off between allocating too large of an array up
front vs. the cost of growing an array each time we insert more on
the end of the list.  Its basically lifted from my marks table in
git-fast-import in C Git, but ported to Java's List interface.

The RevCommitList has a fast conection to a RevWalk in that it can
populate itself by pumping the Generator directly, saving a method
dispatch in between the RevWalk and its clients.  We fill directly
into the tail block until it is full, and then rotate the blocks.
The block rotation to get to the next tail could be more efficient
but we pay this tail block lookup penalty only once per 256 commits
appended into our list, so its not unreasonable.

The list is designed for GUI applications, where a GUI might want to
use this a backing model for a table widget.  The list can be filled
only up to a high water mark, which would be the largest row visible
in the UI.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>