Rewrite the Git history page in terms of revwalk, treewalk
Now that we have a solid revision machinary implementation we can
use that as the backend for the Eclipse history page.  Its quite a
bit faster (~500 ms vs. 160+ seconds) and is able to show more than
one path at a time when multiple paths are selected from the resource
navigator or package explorer.

We also now show the commit message with a user configured font, and
show that font configuration as part of the Appearance settings within
the workspace preferences.

The compare viewer support is very crude, as I'm not fully implementing
enough of the Eclipse APIs yet.  We can reasonably view a difference of
a file so it is at least semi-functional.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>