Allow RevFlag instances to remember which RevWalk created them
It can be handy to track back from a RevFlag to the RevWalk that
created that flag.  Since flags are really mapped down to a 32
bit integer we have a cap of 32 flags per RevWalk, but that is
reduced by the number of core flags that RevWalk reserves for use
by itself and its Generator implementations, so we are really down
to 26 flags available to application level code using RevWalk.  As
such adding in the additional RevWalk pointer to the flag will not
be a huge impact on memory usage, even if most code does not use
the reference.  With only 26 flags per RevWalk the maximum wasted
memory is 104 bytes, which is significantly shorter than the bytes
I have wasted explaining why its OK to waste that amount of memory.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>