Build support for commit searching/highlighting into RevCommitList
Some applications built on top of RevCommitList (e.g. PlotCommitList
and its downstream users) will want to be able to match commits by
a RevFilter graph and flag them for special display (such as drawing
the message in bold so it stands out).  They will also want to skip
forward and backwards through the list to navigate through the commits
the filter matched on.

This set of APIs allows application code to apply a RevFilter graph to
commits to toggle a specific RevFlag on/off.  The RevFlag can then be
efficiently tested for at render time (to choose the correct display)
and also to search forwards/backwards through the list for matches.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>