Allow users to control the font we display our commit history in
This is a new preference to control the font used within the
revision history graph table.  As part of this change we also
now take over full control of drawing the author and date text,
as the font employed may differ later on based upon on a filter
set through some sort of filter interface.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>