Tweak the size of the commit dots in the visualizer
Not exceeding LANE_WIDTH - 2 ensures we leave a little bit of space
around the dot in the case of the row height being larger than the
lane width, due to say a very large font set by the user.  I also
changed the default to be 6 smaller than the height of the row as
it looks a little better to have a bit more gap between the dots
on a standard row height.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>