Use post SelectionListener to trigger display of commit message and files
A normal SelectionListener triggers on each keystroke if you use the
down arrow or up arrow to navigate in the history page's commit graph.
This can get really slow as you have to wait for the viewer to redraw
the message and the file list before you can move to the next commit.

By installing our trigger as a post selection listener we wait until
all UI events have gone idle, and then the update happens.  This way
we are reasonably certain the user has settled on a commit that they
want to see before we waste their time getting its details.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>