Null out the history page refresh job when it completes
Currently the job contains a reference to the full list of commits.
We may in the future be able to GC that list away sooner, but we
cannot do that if our job is still pinned in the history page.  So
clearing out the job reference is easy enough with a job listener
and helps to avoid such leaks later on in life.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>