Define ObjectId.copy() to produce an immutable copy of this id value
In places where we heavily extended ObjectId with lots of additional
data applications may want to shed that data yet retain the unique
identity value of a particular subclass instance.  Since the ObjectId
is that unique identity we offer a form of clone that copies off the
identity information only and stores it into the smallest container
available (ObjectId itself).

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>