Abstract out the job cancellation in our history page and document it
Why we need to ditch our walker when we cancel a job may not be
obvious to the reader, as it has to do with thread-safety of the
walk instance and its data structures.  Since its a performance
sink-hole to be ditching the walk unnecessarily, but it really is
necessary here for thread-safety, we need to explain it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>