Limit the number of UI refreshes the history page job causes
We spend most of our CPU time generating out the first block, but
it often pops through in .5s - 1.5s, depending on the path limiter
being used.  Once that block is obtained we shouldn't bother with
a UI update unless we have spent at least 250 ms since we posted
the last UI update.  Anything that is more frequent than 250 ms
is too small to really be noticed by the end user and is just a
a waste CPU cycles with unnecessary SWT table changes.

Since we use a TOPO sort we typically can get the remainder of the
list (after the first block) in under 250 ms, as it is just moving
the data from the TOPO sort buffer out to the final RevCommitList
we use for plotting, along with the small additional overhead put
in place by the plotting itself.  Therefore it is likely that the
remaining blocks will pop out in the 250 ms window, and that is
exactly what we are seeing in a 12,996 commit history from git.git.

We also avoid updating the table after the last block if we did the
update previously when the block ended normally and there were no
more commits in the TOPO sort buffer.  This is a rare case but is
possible if the result happens to be even divisible by 256.

Finally I noticed during testing that our block sizes were 257, not
the 256 we intended to use.  We really want 256 as it matches up to
the RevObjectList block sizes and thus each iteration through the
main loop of our job causes exactly one block allocation before we
check for a UI update.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>