Override getInput in the history page to conform to Eclipse contract
We replace the real IResource given to us with our own ResourceList,
allowing the history page to show multiple resources at once within
the same graph.  This works great until some part of Eclipse calls
our view and tries to get the current IResource input, to see if it
needs to create a new view or can reuse the existing one.  We now
unpack our ResourceList and return the single IResource within it
if there is only one IResource in the list.  This should make us
better conform to the IHistoryPage API.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>