Don't plot lines to parents who have no lane assigned
If the parent commit does not have a lane assigned yet then there are
two possible reasons why it is lacking such an assignment:

 - The parent has not yet been plotted, but will be soon.  Until then
   we cannot hope to render a line to it, as we do not know where it
   will appear.  Not drawing the line is an easier option.

 - The parent doesn't appear in the graph as a node itself but it was
   not culled by the revision machinary during rewrites.

The latter case can happen when using a path limiter.  At present we
get a horribly disconnected graph for the revision history of git.git
revision.c when starting from 0e545f75169e, and I suspect it is due
to the path limiter not quite producing the correct rewrites.  I am
seeing different parents in gitk vs. egit for 4d1012c370.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>