Use FIFO order for pending queue during RevWalk with no sort order
Using LIFO order here is nuts.  We want either FIFO behavior or
commit timestamp behavior.  The commit timestamp is a close estimate
to TOPO, which ensures all children emit before any parent commit.
In practice many projects have commit timestamps ordered such that
there are very few ties or out of order events caused by clock skew.

When using UINTERESTING to flag commits that we don't want we need
to make sure the commits carry through to the parents at a roughly
even rate, so that the uninteresting side branch reaches the merge
base at the same time as the interesting one we are outputting.
If the interesting one runs in LIFO order it could get ahead of
the uninteresting side branch and produce commits we do not want.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>