Enable automatic word wrapping in the history page commetn viewer
The older history page that I replaced supported line wrapping the
comment viewer, and offered a preference in the view menu to allow
the user to enable or disable the comment line wrapping.  Folks who
work on egit really want the line wrapping feature to be available
so here it is again.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>