Split history page's lower pane into 70/30 by default and save layout
The history pane's comment area is usually more interesting then
the file name area, so we want to dedicate more screen space to
the comment area when possible.

We now also save the split positions as preferences, so they will
automatically restore back to the same layout on the next view of
the history page.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>