Support copy operations from any of the history page areas
We now support the global workbench actions of "Copy" and "Select All"
for any of the 3 major controls of the history page.  Copying from the
commit DAG copies the SHA-1 of that commit; copying from the commit
message viewer copies the selected part of the message; copying from
the file list copies the file path names.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>