Create a context menu on the history page
We now offer our standard copy, select all and view control actions as
part of the context menu for the history page.  This allows users to
right click on the view and control things just as easily as if they
used the main menu bar, keyboard shortcuts or the view's private menu
from its toolbar.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>