Allow the history refresh job to update the UI only once per second
Once we have our first batch of results that is 256 rows of data the
user can see it in the results table, and probably needs to scroll to
view all of those rows.  Typically we can pop the rest of the results
through before the user has had a chance to digest the table, or start
to guess how long the results are.  Since delaying UI update events
really improves the overall throughput reducing our update rate to
only once every other second makes the results nearly instant even for
a repository about the size of git.git.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>