Remove the old history page and topology walker
Now that we have treewalk and revwalk available in jgit we do not need
the prior walker implementation put together by Robin.  Both of the new
walkers are faster and more flexible, as well as being easier to use,
so we do not want to promote usage of the old walker implementation.

To fully remove the old walker implementation we also had to redo the
file history provider implementation, making it more accurately match
the git theory of operation.  We also fixed quite a bit of code that
was just broken, such as the working tree revision not actually being
able to read the working tree data stream.

At 6811 lines deleted and 807 lines inserted this is a fairly large
change all thrown into a single commit, but its worth the pain to
get the tree pruned down to only live code.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>