Define two TreeWalk utility factory methods for single path lookup
If an application just needs to resolve a single path in a tree it
can be faster to use a one-shot TreeWalk with the correct path in
a filter then doing the lookup through the Tree/TreeEntry API, as
the Tree API requires lots of object allocations for paths we do
not want to access.

These utility functions allow construction of a one-shot tree walk
for either a group of trees or a single RevTree, and position the
walker at the one requested entry.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>