Rename RepositoryMapping.recomputeMerge to fireRepositoryChanged
Now that we no longer maintain a standard git tree as our index
and we no longer have a merge tree laying around to tell us the
dirty state of each file we aren't really recomputing a merge here.
This method is really just firing a listener so the UI plugin can
schedule an update of the resource decorator.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>