Remove unnecessary traces from resource decorator
These trace statements are slowing down our resource decorator, which
may have a lot of work to perform if there are many files visible in
the explorer tree at once.  Even if the user has logging turned off
we are still constructing these strings, which hurts performance.

Most of this decorator code is considered to be stable, and if we do
want to see what is going on here we can always use a debugger from a
host workbench or insert a handful of trace lines at key points.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>