Break out the Java 6 executable support from GitIndex for reuse
This code to support access to the executable flag is handy for
locations where we need such functionality but don't want to
require building with Java 6 to setup the method calls nor do we
want to dirty everyone with ugly reflection support.

On Win32 the executable flag is basically useless so we do not
ever support it there, even if the Java 6 JRE is being used and
the API is available.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>