Support Cygwin path translations when running on Win32 with Cygwin
If the user has cygpath.exe in their PATH environment variable that
spawned the JRE we may need to translate paths from the Cygwin UNIX
style into the native Windows paths so that the native Win32 JRE is
able to read them.

This allows jgit to read repositories that are created by a Cygwin
C Git, or that uses Cygwin style symlinks to setup a workdir (where
.git/refs and .git/objects are symlinks to the real repository).

This support isn't quite complete as the ".git/logs/refs" directory
is usually also a symlink, but we don't first resolve it into the
Windows native format.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>