Make AbstractRevQueue actually extend Generator
The Generator API and the RevQueue APIs are very close; about the
only difference between them was the use of "pop" vs. "next" and
the outputType being supported on the Generator API.  By putting
them together into a single type hierarchy we can treat all of our
queue types as Generators and avoid an extra level of indirection
when we are forced to buffer commits.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>