Enable Mac OS X specific options to 'brand' JGit
By using these Apple specific options on Mac OS 10.4 we can get
the application title to be not the name of our class, but instead
a string of our choosing.  This makes it fit into the desktop a
tiny bit better than showing a long class name.

Some of this may have changed with Mac OS 10.5, due to an improved
Aqua implementation of AWT being deployed in that release.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>