Decrease the commit dot size for SWT and AWT renderers
The default dot size was coming out too large on SWT; it took up
most of the entire cell and looked very out of place relative to
the lines connecting the dots.  We now use 50% of the smallest side
of the cell (row height or lane width).  Odd sized dots are rounded
up to an even dot size.

The AWT renderer now centers the lines in the middle of the stroke,
as it was placing them out of position relative to the dots.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>