Change history page table to SWT.VIRTUAL.
It makes the history page show about the same speed as gitk on my
eclipse.

From the eclipse API:

"Style VIRTUAL is used to create a Table whose TableItems are to be
populated by the client on an on-demand basis instead of up-front.
This can provide significant performance improvements for tables
that are very large or for which TableItem population is expensive
(for example, retrieving values from an external source)."

Signed-off-by: Roger C. Soares <rogersoares@intelinet.com.br>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>