Allow RefLock users to write directly to the output stream
When updating the index we want to use the same locking strategy
that we use when writing to ref files.  Its the same one that is
used within C Git, and therefore is the only safe way to do IO on
the same repository as C Git.

We now export a custom java.io.OutputStream that can write into the
temporary file and then release the lock channel when the output is
closed by the caller.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>