Catch any Errors during lock writes and automatically unlock
We already catch RuntimeException and IOException while writing to
a lock file and translate those into an unlock followed by a throw
of the original exception.  This prevents orphaned locks on disk as
much as possible, even though the JVM may still be running and thus
has not cleaned up at the OS level.

We now also catch Error and rethrow it, ensuring we still try to
unlock the file when faced with an OutOfMemoryError while dumping
the ObjectId to the output stream.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>