Rename RefLock to LockFile as its a generic locking system
This locking strategy is suitable for a lot more formats than just
refs.  In C Git it gets used for the dircache (aka .git/index), for
a symbolic ref (aka .git/HEAD) and also for the packed-refs file.
Consequently it is a very commonly employed strategy that we also
need to use throughout jgit to remain as compatible as possible with
C Git and its on disk repository format.

To aid confusion our LockFile depends upon java.nio.channels.FileLock.
This is something we inherited from RefLock and its desire to use the
platform locking strategy available from the JRE.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>