Teach RawParseUtil how to format base 10 numbers it reads
In some file formats we need to be able to efficiently produce a
base 10 numeric in a byte[], much as how we need to parse those
through RawParseUtil.parseBase10.

For efficiency reasons we write backwards into the passed buffer,
making it necessary for the caller to produce the output in the
reverse order that it would parse the input with.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>