Teach FileMode what the proper object type is for each tree entry
Every file mode has a specific type of object that it expects the
associated SHA-1 to be referencing when the mode is used for one of
the entries of a Git tree or index (aka dircache).   Having this in
the FileMode object constant makes application code easier to follow
when we are talking about displaying a tree or verifying it is valid.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>