FindToolbar port to the new history page.
This is a port from the find toolbar to the new history page.
It is replacing the current highlight flag in order to have the
find feature back but most of the code from the toolbar shall be
replaced to make use of the new infrastructure added into jgit
(RevFlag and RevFilter).

Signed-off-by: Roger C. Soares <rogersoares@intelinet.com.br>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>