Change the find toolbar find-in prefs into radio rotating options.
The find toolbar now searches only in one field at a time and the user
can change the find-in field by clicking on the button. The button shows
an icon so the user knows at all times what field is selected.

We don't have the option to find in 2 or more fields at the same time
anymore, but from day to day usage I noticed that I usually only want to
search in one field at a time. It feels a lot better to me and makes it
more like gitk.

Signed-off-by: Roger C. Soares <rogersoares@intelinet.com.br>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>