Use MutableObjectId in IndexPack to avoid violating encapsulation
The representation of an ObjectId is meant to be opaque to
application level code such as IndexPack.  We shouldn't be
doing 5 word decodes at this level fo the system, but rather
should dump our MessageDigest result into a MutableObjectId
(which does the transform for us) and then copy that into
the final destination.

This has roughly the same performance (no allocations) but
does store the temporary values into instance members and
not on the stack.  This is a very minor performance penalty
(if any) for the improved encapsulation.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>