Use common NB.decodeUInt32 in IndexPack
The IndexPack code was developed before NB came along, so it copied
the entire readUInt32 from someplace else.  Really we want to just
use the common version in NB as its slightly faster on some systems
(better machine code comes out of the JIT) and its just less code
we are hauling around and maintaining ourselves.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>