Replace IndexPack writeUInt32 with NB.encodeInt32
Since we already have a common function to write a 32 bit
unsigned (or signed) network byte order integer we do not
need to create our own in IndexPack.  Better to reuse the
code we already have.

Since rawoe is only used for the ObjectId values within
this loop we can instead pack it out an extra 4 spaces
and fill the entire index record into one byte array.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>