Don't allow IndexPack to create index v1 for larger than 4 GB of data
A version 1 pack index constructed around a packfile that has an
object offset past the 4 GB boundary will not be readable by jgit
or by C Git.  To access such a pack a v2 index must be used, as
the 64 bit offset table is needed to link to these later objects.

Rather than build a possibly corrupt index when we round down the
64 bit position to a 32 bit position we check for any possible
overflow and throw an exception if we would create a bad index.
In the future we should automatically switch over to generate a
v2 index when overflow on position would happen, or if the user
has configured v2 indexes as their preference for this (or any)
repository.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>