Teach RepositoryConfig to expose multi-value keys to callers
Some parts of the git config file (e.g. "remote.origin.url") can
have more than one value within the configuration file, and the of
the entries may be relevant to the caller.  In a case like this
we want to obtain all of the values as a String[] so they can be
enumerated in application level code.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>