Remove unused RemoteSpec now that RemoteConfig, RefSpec superseded it
The RemoteSpec class had good intentions of handling the remote
configuration data necessary for a fetch or push operation, but
wasn't quite abstracted out in a way that would lead to good reuse,
nor is it actually used by the existing fetch implementation.

Now that RemoteConfig and RefSpec accurately model the remote
block of a config file we won't be needing this first draft
attempt anymore.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>