Allow RevCommit callers to access the parent array directly
Some application level code can be written easier using a Java 5
for loop style syntax, which requires an array or an Iterateable.
Rather than using ints to loop over the parents we can let the
application get to the internal array, and iterate over that.

Since we already expose the byte[] for the complete body there
is little risk to also exposing the complete parent array.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>