Define an IndexPack utility to generate a temporary base name
When using IndexPack on the receiving end of a network connection
(e.g. during fetch to get the data from the remote, or if we were
to implement the server side of the push protocol) it is necessary
to generate a temporary file name that will hold the pack and its
index prior to being rotated into the objects/pack directory for
final storage.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>