Teach IndexPack how to fix thin packs received from the network
Network transfers can use "thin packs", packs where base objects
for one or more of the contained deltas are not part of the pack.
This permits two peers to exchange objects with the minimum amount
of bandwidth necessary, but the recipient must add the missing base
objects by copying them from the local repository into the pack.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>