Strip DOS CRLF braindamage from recent jgit class additions
Sadly I added these to the repository from a system that uses CRLF
as its default line ending.  The only true form for this project
should be LF, as most developers are on UNIX like systems, and the
Git standard is to use LF in the repository and convert back and
forth during checkout/staging.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>