Teach URIish to recognize git+ssh and ssh+git synonyms for ssh
Cogito (which is now dead) has made "git+ssh://" and "ssh+git://"
style URIs popular among a number of users and many Git tutorials
have been written using these sorts of URIs.  This is not the
preferred way to denote an SSH connection, but it is handled by C
Git and is in use by a number of users.  We need to support those
in jgit.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>