Only parse a RevCommit's parents once
If we have disposed of a RevCommit we only have tossed its byte
buffer holding the full message body.  The parents stay in our memory
space so that we can traverse them during execution.  If the commit
is using any sort of parent rewriting we do not want to lose that
if the caller forces us to reparse to recover the message body.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>