Allow applications to ask RevWalk to carry flags onto parents
A very common idiom for RevWalk based applications is to mark
a number of commits with some RevFlag or two, and then walk the
history, carrying those flags onto the parents as we go.  At some
point we are looking for a combination of flags, or just need the
flags there for future reference.

We now permit applications to specify RevFlags which we will carry
onto the parents during a normal walk, allowing the application to
free itself from this part of the process.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>