Allow RevWalk users to ask for FIFO style ordering
Some applications just want (or require) a FIFO style ordering
when we walk commits.  This is easier for us than date order,
and is what we use by default in the queue when the sort order
is NONE, but we shouldn't promise that.  Instead we use a new
RevSort value to enforce FIFO requirements.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>