Report progress on console if operation takes more than 500 ms
Rather than starting our timer on each task, we start the timer for
the overall operation, and instead of using a timeout of 1000 ms for
the first output we use 500 ms.  In many cases jgit is fast enough
that we almost never see any sort of progress from it, even though
the same operation in C Git shows progress.  Our timer was just a
tad too long and can leave the user wondering what we are doing.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>